==> scripts/34_snp_rr_vs_ctrl.py <==
#testing differences in snps genotypes between
#controls and RR patients using pearson's
#chi-squared test.

import pandas as pd
from scipy import stats

#reading list of available snps

with open('./data/snp/snp_list.csv', 'r') as f:
	snp_list = f.readline().strip().split(',')

#reading table containing snps genotype frequencies
#in european population

snp_df = pd.read_csv('./data/snp/snp_hw.tsv', sep = '\t', header = 0, index_col = 0)

#for each snp, extracting frequencies of 3 possible genotypes

euro_gen_freq = {} #frequencies I will use as expected in tests

for snp in snp_list:
	pp = list(snp_df[snp_df['id'] == snp]['pp'])[0]
	pq = list(snp_df[snp_df['id'] == snp]['pq'])[0]
	qq = list(snp_df[snp_df['id'] == snp]['qq'])[0]
	euro_gen_freq[snp] = [pp, pq, qq]

#WHERE:
#	pp is equal to omozygote for most frequent allele
#	pq is eterozygote
#	qq is omozygote for least frequent allele

#now I will create a list of all 3 possible genotypes for each snp
	#this will contains the 2 alleles for each snp
snp_alleles = []
for snp in snp_list:
	entry = []
	entry.append(snp)
	entry.append(list(snp_df[snp_df['id'] == snp]['p'])[0])
	entry.append(list(snp_df[snp_df['id'] == snp]['q'])[0])
	snp_alleles.append(tuple(entry))
	#this dictionary will contain the possible genotypes.
	#since I cannot know wether the eterozygotes are encoded
	#as pq or qp, I'll have to add both the formattings
snp_genotp = {}
for entry in snp_alleles:
	omop = entry[1] + entry[1]
	eth1 = entry[1] + entry[2]
	eth2 = entry[2] + entry[1]
	omoq = entry[2] + entry[2]
	snp_genotp[entry[0]] = (omop, eth1, eth2, omoq)

#adding a check, just to see if this is running :)
for entry in snp_genotp:
	print(f'{entry} --> {snp_genotp[entry]}')

#now I'm going to read the data about snps from the 'database'

df = pd.read_excel('../database/db.xlsx', sheet_name = 'SM NM', usecols = 'A,CB,JI:PW')
#dropping control who is messing up all over the place :¬|
df = df[df['Codice PLASMA NM'] != 'C16']
#dropping empty columns
df.dropna(axis = 1, how = 'all', inplace = True)
#renaming columns because comfort
snp_list.insert(0, 'patient_id')
snp_list.insert(1, 'class_int')
df.columns = snp_list
snp_list.pop(0)
snp_list.pop(0)
#checking again
print(df)

#Isolating controls

ctrl_df = df[(df['class_int'] == 6) | (df['class_int'] == 7)]
print(ctrl_df)

#Isolating RR patients

rr_df = df[df['class_int'] == 3]
print(rr_df)

#now I will proceed to recode the genotypes using integers.
#In particular, pp genotype will be 0, pq 1 and qq 2.
#This will be useful for testing, clustering and other stuff.

ctrl_samp = {}
for entry in snp_genotp:
	samp = []
	snp_row = list(ctrl_df[entry].dropna())
	recode = []
	for gen in snp_row:
		if gen == snp_genotp[entry][0]:
			recode.append(0)
		elif gen == snp_genotp[entry][1] or gen == snp_genotp[entry][2]:
			recode.append(1)
		elif gen == snp_genotp[entry][3]:
			recode.append(2)
	ctrl_samp[entry] = recode

rr_samp = {}
for entry in snp_genotp:
	samp = []
	snp_row = list(rr_df[entry].dropna())
	recode = []
	for gen in snp_row:
		if gen == snp_genotp[entry][0]:
			recode.append(0)
		elif gen == snp_genotp[entry][1] or gen == snp_genotp[entry][2]:
			recode.append(1)
		elif gen == snp_genotp[entry][3]:
			recode.append(2)
	rr_samp[entry] = recode

#turns out the genotypes for some of the snps in our db
#do not match the 'possible' alleles reported on ncbi.
#Since I trust NCBI way more than our data, I will ignore
#those snps and move on only with the 'good' ones

clean_ctrl = dict(ctrl_samp)
clean_rr = dict(rr_samp)
for snp in ctrl_samp:
	if len(ctrl_samp[snp]) == 0:
		del clean_ctrl[snp]
		del clean_rr[snp]
for snp in clean_rr:
	print(snp, clean_ctrl[snp], len(clean_ctrl[snp]))
for snp in clean_ctrl:
	print(snp, clean_rr[snp], len(clean_rr[snp]))

#and now getting a list of the remaining snps

snps_clean = list(clean_ctrl.keys())

#since I don't have expected frequencies this time,
#I will estimate them on RR patients. Then I'll compare
#the expected frequencies and test them against
#the observed frequencies in controls. In this way
#I can check wether or not they follow the same
#distribution

rr_freq = {}

for snp in snps_clean:
	tot = len(clean_rr[snp])
	if tot != 0:
		pp = clean_rr[snp].count(0)/tot
		pq = clean_rr[snp].count(1)/tot
		qq = clean_rr[snp].count(2)/tot
		rr_freq[snp] = [pp, pq, qq]
	else:
		rr_freq[snp] = [0, 0, 0]

#now I can organize the data for the chi-square tests:

chisq_data = {}
for snp in snps_clean:
	pp = clean_ctrl[snp].count(0)
	pq = clean_ctrl[snp].count(1)
	qq = clean_ctrl[snp].count(2)
	pp_exp = rr_freq[snp][0]*len(clean_ctrl[snp])
	pq_exp = rr_freq[snp][1]*len(clean_ctrl[snp])
	qq_exp = rr_freq[snp][2]*len(clean_ctrl[snp])
	chisq_data[snp] = [[pp, pq, qq], [pp_exp, pq_exp, qq_exp], len(clean_rr[snp])]

#now testing

f = open('./data/snp/tests/ctrl_vs_rr.tsv', 'w')
f.write('snp_id\tsamp_dim\tChi-stat\tp-val\n')

for snp in snps_clean:
	if((chisq_data[snp][2] != 0) and (0 not in chisq_data[snp][1])):
		stat = stats.chisquare(f_obs = chisq_data[snp][0], f_exp = chisq_data[snp][1])
		print(chisq_data[snp][0], chisq_data[snp][1])
		f.write('{}\t{}\t{}\t{}\n'.format(snp, chisq_data[snp][2], stat[0], stat[1]))
f.close()

#applying benjamini-hochberg correction to p-values

def benjamini_hochberg(dataframe, dataframe_column):
	#dataframe_column is the name or index of the col containing p-values
	dataframe.sort_values(by = dataframe_column, inplace = True)
	p_values = list(dataframe[dataframe_column].astype(float))
	q_values = []
	n = len(p_values)
	for i, p_value in enumerate(p_values):
		q_values.append((p_value * n)/(i+1))
	#adding q_values column to dataframe
	dataframe['q_values'] = q_values
	dataframe.sort_values('q_values', inplace = True)
	return dataframe

df = pd.read_csv('./data/snp/tests/ctrl_vs_rr.tsv', sep = '\t')
df_corr = benjamini_hochberg(df, 'p-val')
df_corr.to_csv('./data/snp/tests/ctrl_vs_rr_bh.tsv', sep = '\t')
print(df_corr)
